safari-600.7-branch/LayoutTests/storage/indexeddb/factory-basics-workers-expected.txt

[Worker] Test the basics of IndexedDB's IDBFactory.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/factory-basics.js
[Worker] indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;
[Worker] 
PASS [Worker] typeof indexedDB.open is "function"
PASS [Worker] typeof indexedDB.cmp is "function"
PASS [Worker] typeof indexedDB.deleteDatabase is "function"
PASS [Worker] typeof indexedDB.webkitGetDatabaseNames is "function"
PASS [Worker] typeof indexedDB.getDatabaseNames is "undefined"
[Worker] request = indexedDB.webkitGetDatabaseNames()
PASS [Worker] request.source is null
[Worker] databaseNames = event.target.result
PASS [Worker] databaseNames.contains('storage/indexeddb/factory-basics') is false
PASS [Worker] databaseNames.contains('DATABASE THAT DOES NOT EXIST') is false
[Worker] request = indexedDB.open(name)
PASS [Worker] request.source is null
[Worker] event.target.result.close()
[Worker] indexedDB.webkitGetDatabaseNames()
[Worker] databaseNames = event.target.result
PASS [Worker] databaseNames.contains('storage/indexeddb/factory-basics') is true
PASS [Worker] databaseNames.contains('DATABASE THAT DOES NOT EXIST') is false
[Worker] request = indexedDB.deleteDatabase('storage/indexeddb/factory-basics')
PASS [Worker] request.source is null
[Worker] indexedDB.webkitGetDatabaseNames()
[Worker] databaseNames = event.target.result
PASS [Worker] databaseNames.contains('storage/indexeddb/factory-basics') is false
PASS [Worker] databaseNames.contains('DATABASE THAT DOES NOT EXIST') is false
[Worker] indexedDB.deleteDatabase('DATABASE THAT DOES NOT EXIST')
PASS successfullyParsed is true

TEST COMPLETE
